Mythmusic: Suppress some compiler warnings. (cherry picked from commit 6e71eda7fab37f8c657d357979ff051ccf324afc)

+++ mythplugins/mythmusic/mythmusic/visualize.cpp
@@ -112,14 +112,14 @@ void Spectrum::resize(const QSize &newsize)
     rects.resize( scale.range() );
     unsigned int i = 0;
     int w = 0;
-    for (; i < rects.size(); i++, w += analyzerBarWidth)
+    for (; i < (uint)rects.size(); i++, w += analyzerBarWidth)
     {
         rects[i].setRect(w, size.height() / 2, analyzerBarWidth - 1, 1);
     }
 
     unsigned int os = magnitudes.size();
     magnitudes.resize( scale.range() * 2 );
-    for (; os < magnitudes.size(); os++)
+    for (; os < (uint)magnitudes.size(); os++)
     {
         magnitudes[os] = 0.0;
     }
@@ -164,7 +164,7 @@ bool Spectrum::process(VisualNode *node)
 #endif
 
     index = 1;
-    for (i = 0; i < rects.size(); i++, w += analyzerBarWidth)
+    for (i = 0; i < (uint)rects.size(); i++, w += analyzerBarWidth)
     {        
 #ifdef FFTW3_SUPPORT
         magL = (log(sq(real(lout[index])) + sq(real(lout[FFTW_N - index]))) - 22.0) * 
@@ -255,7 +255,7 @@ bool Spectrum::draw(QPainter *p, const QColor &back)
     double r, g, b, per;
 
     p->fillRect(0, 0, size.width(), size.height(), back);
-    for (uint i = 0; i < rects.size(); i++)
+    for (uint i = 0; i < (uint)rects.size(); i++)
     {
         per = double( rectsp[i].height() - 2 ) / double( size.height() );
 
@@ -569,7 +569,7 @@ bool Squares::draw(QPainter *p, const QColor &back)
 
 #if defined(FFTW3_SUPPORT) || defined(FFTW2_SUPPORT)
     QRect *rectsp = rects.data();
-    for (uint i = 0; i < rects.size(); i++)
+    for (uint i = 0; i < (uint)rects.size(); i++)
         drawRect(p, &(rectsp[i]), i, center, w, h);
 
 #else
@@ -809,14 +809,14 @@ void Gears::resize(const QSize &newsize)
 
     rects.resize( scale.range() );
     int i = 0, w = 0;
-    for (; (unsigned) i < rects.size(); i++, w += analyzerBarWidth)
+    for (; (unsigned) i < (uint)rects.size(); i++, w += analyzerBarWidth)
     {
         rects[i].setRect(w, size.height() / 2, analyzerBarWidth - 1, 1);
     }
 
     int os = magnitudes.size();
     magnitudes.resize( scale.range() * 2 );
-    for (; (unsigned) os < magnitudes.size(); os++)
+    for (; (unsigned) os < (uint)magnitudes.size(); os++)
     {
         magnitudes[os] = 0.0;
     }
@@ -860,7 +860,7 @@ bool Gears::process(VisualNode *node)
     rfftw_one(plan, rin, rout);
 #endif
     index = 1;
-    for (i = 0; i < rects.size(); i++, w += analyzerBarWidth)
+    for (i = 0; i < (uint)rects.size(); i++, w += analyzerBarWidth)
     {
 #ifdef FFTW3_SUPPORT
         magL = (log(sq(real(lout[index])) + sq(real(lout[FFTW_N - index]))) - 22.0) *
